Use the 'snapshot' version of Commons-CLI
Specifically we'll need this feature from the upcoming v1.10 of commons-cli:
https://github.com/apache/commons-cli/pull/334

The assumption here is that the release should be available *reasonably soon*. No release date has been announced, but going by the pace of previous releases and the activity/minor version diff it seems .
It was either this or patch the class implementation temporarily, doing this seemed like the better option.
Collaborators/testers need to obtain a prebuilt package or build from source in the meantime.